backend/requirements.txt
# Dev server: serves frontend + data + Gemini proxy (Ask AI)
flask>=3.0
httpx>=0.27
python-dotenv>=1.0
gunicorn>=22.0
pgeocode>=0.5
